Fixed writing to registers B & C, and added basic support for interrupts
I noticed that the instruction DEC B wasn't executed correctly. The
problem was a mistake where writing to register B and C was swapped.

I have also added basic support for interrupts. The IME (Interrupt
Master Enable) is now implemented but does not yet function with
the rest of the code. In addition opcode DI & EI are now supported.