Moved build system from GNU Makefile to CMake and moved header files to include dir
I struggled quite hard to make a Makefile that was also usable for
a testing framework. What most testing frameworks have in common is
that they use CMake. Therefore I created a basic CMakeLists.txt
file that compiles my project the same way is did as with the
Makefile.

I also moved all header files to the include directory because it
looks more cleaner for when I will create a new subdirectory called
tests that will include the header files. Previously all headers
files related to my source files where in the same directory
but I did not want to include the source directory.