Fixed background and window tile selection
The tile map has size 32x32. I flatten this to an array with size
1024. The function which maps a pixel on screen to an index in
this flattend array, returns a value with type u8. 8 bits is not
enough information to index the entire array. I changed the return
type to int.

This results in Blargg test to be rendered correctly. However
Tetris is still broken.